+++ booking_web/src/components/pages/UpdateBooking.jsx
@@ -1,49 +1,16 @@
-import React, { useEffect, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import React, { useEffect } from "react";
 import {Box} from "@chakra-ui/react";
-import { useSetRecoilState, useRecoilState } from "recoil";
+import { useSetRecoilState } from "recoil";
 
-import { getAllBooking, updateBooking } from "../../utils/api";
 import { optionDispState } from "../../atom/state";
-import { useMessage } from "../../hooks/useMessage";
 import { MainHeading } from "../atoms/heading/MainHeading";
-import { SelectMember } from "../molecules/SelectMember";
-import { InputDate } from "../atoms/input/InputDate";
-import { nameState, dateState, hourState, minuteState, numberOfAdultsState, numberOfChildrenState, bookingCategoryIdState, noteState } from "../../atom/state";
-import { RadioBookingCategory } from "../molecules/RadioBookingCategory";
-import { TextNote } from "../molecules/TextNote";
 import { RegisterBookingButton } from "../atoms/button/RegisterBookingButton";
 import { InputArea } from "../organisms/layout/InputArea";
+import { useFetchBooking } from "../../hooks/useFetchBooking";
 
 export const UpdateBooking = () => {
-  const { id } = useParams();
   const setOptionDisp = useSetRecoilState(optionDispState);
-  const setDate = useSetRecoilState(dateState);
-  const setHour = useSetRecoilState(hourState);
-  const setMinute = useSetRecoilState(minuteState);
-  const setName = useSetRecoilState(nameState);
-  const setNumberOfAdults = useSetRecoilState(numberOfAdultsState);
-  const setNumberOfChildren = useSetRecoilState(numberOfChildrenState);
-  const setBookingCategoryId = useSetRecoilState(bookingCategoryIdState);
-  const setNote = useSetRecoilState(noteState);
-
-
-  const fetchBooking = async () => {
-    const allBooking = await getAllBooking();
-    const booking = allBooking.find((booking) => booking.id === Number(id));
-    setDate(booking.date);
-    setHour(booking.time.slice(0, 2));
-    setMinute(booking.time.slice(3, 5));
-    setName(booking.name);
-    setNumberOfAdults(booking.number_of_adults);
-    setNumberOfChildren(booking.number_of_children);
-    if (booking.booking_category.name === "LINE") {
-      setBookingCategoryId("1");
-    }else {
-      setBookingCategoryId("2");
-    }
-    setNote(booking.note);
-  };
+  const { fetchBooking } = useFetchBooking();
 
   useEffect(() => {
     fetchBooking();
